Always push a NULL to the end of parser.nodes
Not pushing a NULL to the end of parser.nodes created
undefined behavior, because the last part of parser.nodes
was uninitialized.
Also, having exactly 32 nodes in parser.nodes, would previously segfault
because it would overflow when iterating the nodes, because it would not have
a NULL terminator, or in this case an unititialized terminator, which is now also fixed.