Add and + or operators
To use the 'and' operator you need to do:
expr1 & expr2

The 'and' operator returns 1 if both statements are true

To use the 'or' operator you need to do:
expr1 | expr2

The 'or' operator returns 1 if one of the two expressions equal one.

Both operators gradually evaluate the full expressions,
which means they won't evaluate the two expressions and then check
if they are 1 or 0, but will evaluate the first expression and check
if it's 1 or 0, then depending on the result, it will stop or continue
evaluating the second expression.

This means that '0 & blame("error")' wouldn't error,
because the first one is 0, which will make the evaluation stop,
but '1 & blame("error")' will error, because the first expression is 1,
which makes the evaluation continue.

This process is similar to that of other languages like C.