Fix float parsing to actually work
Beforehand we did some bad float parsing math which didn't work,
this was fixed in this commit.

This commit also adds better float printing by changing the precision
of double printing with the `%g` formatter which will allow printing numbers
with more than 3 digits after the decimal point without it not showing up.