Add type casts using the `to` keyword
From now on, you can cast values to other types, for example:
:number "23" to Number
log :number * 10
this will parse a the string 23 as a number and store it into :number, later using
it in a calculation, finally logging 230.
The different casts currently possible are:
Void to Number = 0
Void to String = "Void"
Number to String = the string representation of the number
String to Number = the string parsed as a number
String to Stack = a stack filled with the ascii values of the characters in the string.