Add argv and filename builtin variables
The argv variable stores a stack with the arguments from *after* the
code stream was defined, and the filename variable stores the filename
incase a file was run, and a Void incase a string was run.
To access the argv variable, the :_Argv variable must be used.
To access the filename variable, the :_Filename variable must be used.