Add the abillity of setting and modifying members of values
For example:
:number ?= 6
:number:name ?= "number"
:number:name += "s are cool"
The above code will create a member for the Number typed "number" and will add
a member called "name" to it. After that, it modifies the member by adding a string to it,
making it equal "numbers are call".
This feature will be useful in the future, when our OOP system gets more complex, and we have methods,
but could also be used now to make struct-like values.