Change precedence of parens and fix state overrides
This commit changes the precedence of the parenthesis expression `(expr)`.
Beforehand we apparently couldn't do something like `log (1 + 2):SomeKey`
and we would error expecting a newline.
From now on it is possible to execute that expression.
This commit also fixes some problems with states being overriden when
there are multiple levels of expression parsing.