Add Stack type methods
This commit adds 2 methods to the stack value.

The first method is "pop" and it pops a value off the stack.
Stack:pop takes zero parameters but has an optional parameter saying what index to pop.
The optional parameter's default value is the length of the stack - 1.

The second method is "findIndexOf" and it finds the index of a matching value/pointer in the stack.