Fix UB in parser when having many commands
When writing a relatively long test for the language, I encountered
UB when running the script.
The problem has been in the source code since the beggining and was caused
because of a type error.
Basically, I allocated `(amount) + 1 * size)` instead of allocating `((amount) + 1) * size)`,
which caused a problem because I just allocated `amount + size` each time, which is a smaller size.
This caused the parser to write to unallocated addresses, which caused UB.