Implement module caching and store modules in the interpreter struct
Module caching is a feature that caches module creation, so that
one can not load the same module twice and get a different value.
This also make changes to modules stay.
The overall idea behind the caching is that it reduces allocations
when importing big modules.

load :Math
:Math:Cos ?= routine(:a) {
    ^:Math:Cos()
}
:cos ?= :Math:Cos
load :Math:Cos
log :Math:Cos = :cos %% will be 1