Add catch ... else
From now on, you can catching non-blames with the else keyword.
In other languages, if an expression errors, it can be handled with a catch,
and the code will continue naturally otherwise, but we can't do that because
our blames are not part of a block but are specific values instead.
Instead of that, what we can do is execute the expression and handle each case
seperately.

Until now, if you wanted to handle catching a non-blame, you couldn't do that easily.
This commit makes it possible to handle each case seperately and even get the evaluated value.

The old syntax allowed you to do:
catch 1 / 0 with :errMsg {
    log :errMsg
}

But if you wanted to call a function and didn't know its output, you were in danger of erroring.
For example, a code like:
~ log "string":split(123)
could not have been handled manually.
Now we can:
load :Types
:idx ?= getstring("") to :Types:Number
catch "string" at :idx with :errOrResult {
    log :errOrResult %% some error message
} else {
    log :errOrResult %% some value
}