Add the abillity for any callable to define how many arguments it takes
From now on, callables can define all kinds of different information
about themselves - they can, first of all, define which cfunc they want
to call, but can also give a function that validates the amount of arguments
it takes, including a flag to tell when to call it.
This allows much better control of callables and much more unified
functionality for calling callables.

This shortens the amount of code needed to write in CFuncs and CFuncMethods function declerations,
because it allows them to declare how many arguments they take instead of each
function validating that differently.

This commit also defines a constructor_info member in the type struct,
which gives construction information, like the CFunc, the amount of arguments
etc.
This is used by Type's op_call to know how to call the constructable.