Add :Graphics module
The :Graphics module is an SDL-2 wrapper module for Rael.
It defines a few useful types, functions and constants used for basic
games development or just for experimenting with graphics-related stuff.
There isn't documentation of the module and how it works yet, but in general
it is really close to how SDL does stuff.
Also, he source code is pretty straight-forward so it shouldn't be too hard
to understand how stuff works.