Read file correctly
Previously, due to lack of knowledge of the fread command, we didn't pass
the right parameters. The reading did work, but it was not the valid way of reading
a file.
Instead of passing the size of the read (sizeof(char)) and the length of the file,
we passed in the length of the file and the constant 1, which read the file correctly
but would fail on a char that is not 1 size long.
There's a possibility the read was slower because of this misuse.