Add RaelStream structure to manage file streams better, and add an instance value
The RaelStream structure is a structure that contains info about a given stream,
making it easier to pass it around so that each instance has the right info about
the stream.

The structure contains `base` (char *) which points to the start of the stream,
`cur` (char *) which points to the current stream position, `length` which stores
the length of the stream, `on_heap` which tells us if we need to free stream_base, and
`name` which tells the name of the stream (usually the filename).

Instances are linked structures which store information about every different code run.
We start with the basic instance, storing a RaelStream and some other useful info, and if
we for example want to (in the future) load a different file, we will call
`interpreter_new_instance`, `interpreter_interpret` and `interpreter_delete_instance`.
The instances make it easy to switch the files we are currently running and come back later.

This update is a move forward into file module support and code eval functions.