Add test script like runtests.py written in rael
The file runtests.rael is a new file that behaves just like runtests.py
but is written in Rael. It isn't the best idea to run it after changes to the codebase,
because if you want to test a new change, the script might not like that change and
will behave or misbehave accordingly.
This is just a nice addition, and maybe in the future, with better versioning, we will
be able to run it using an old version of Rael, making it the official script for testing.