Duplicate keys so that we don't deallocate keys in use
I noticed that we were double free()ing keys that were created in different
instances, because they are stored in the ast and when the instance is deleted,
we free the ast.

From now on we call a `strdup` copy on almost every key set, including setting
members.

This might be slower and more memory intensive, but it allows greater control over our
interpreter environment.