Make instructions polymorphised
This means that from now on, all of our instructions inherit from the basic
RaelInstruction structure.
This allows the instruction structures to get information about themselves,
without losing their identity.
For example, if I pass an instruction to a function, I would still be able
to know its type (for example CatchInstruction) and I would also know all
of the basic details about it: how many references it has, its state etc...

This also allows allocating just enough space, a fact that minimizes the amount
of memory some instructions use, as opposed to our unionized struct of instructions
beforehand, that made every instruction be a constant size. Although this barely
reduces the amount of memory used, it is still helpful because the ast memory is
constant throughout the program lifetime, and because it wasn't the main point of
this change, but just a little improvement that makes our language just a bit better
:^).

Other than that, I also added a RaelInstructionType structure which allows specifying
details about a specific instruction. The information is currently only the run function
and delete function, but in the future, depending on whether I'll need it, I'll also
add other information specifiers.

This will make adding instructions in the future easier.

This kind of polymorphism was used for Rael values, and I saw it made some things much
simpler and allowed bigger and more generalized control over our language, so I used
it here.

There are a few more ways to do this kind of polymorphism, but I've used this kind and
liked it, so I just used it again :).