Add support for mmapping files on posix systems instead of mallocating them
Until now, on every system, we would malloc our files to memory, but it isn't always
the fastest and most efficient way of allocating files to memory.
That's why, on unix-based systems, there exists `mmap`, which maps files onto memory
in an efficient way.
In this commit, I made it so that unix-based systems will use `mmap` to allocate Rael
source files, and made it so that other non-unix-based systems will just use `malloc`
as before.