Add simple :Encodings module with base64 encoding support
This commit adds a hand-crafted module that currently has just base64 functions.
In the future I will add more encodings/encryptions.

It did take quite some time for me to think of a base64 encoding/decoding
logic that actually works, although to be fair the encoding is about misaligned bytes
which is pretty weird.